backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
email-validator==2.1.0
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6